CAD_PHASE1
Phase 1: Problem Definition and Design Thinking   

In this part you will need to understand the problem statement and create a document on what have you understood and how will you proceed ahead with solving the problem. Please think on a design and present in form of a document.   

 

Problem Definition: The project involves training a machine learning model using IBM Cloud Watson Studio and deploying it as a web service. The goal is to become proficient in predictive analytics by creating a model that can predict outcomes in real-time. The project encompasses defining the predictive use case, selecting a suitable dataset, training a machine learning model,deploying the model as a web service, and integrating it into applications. 

Design Thinking: 

    Predictive Use Case: Define a use case for predictive analytics, such as predicting customer churn or product demand. 

    Dataset Selection: Choose a relevant dataset to train the machine learning model. 

    Model Training: Select a suitable machine learning algorithm and train the model using IBM Cloud Watson Studio. 

    Model Deployment: Deploy the trained model as a web service using IBM Cloud Watson Studio's deployment capabilities. 

    Integration: Integrate the deployed model into applications or systems to make realtime predictions.