CAD_PHASE2
Phase 2: Innovation

In this phase you need to put your design into innovation to solve the problem. 
Explain in detail the complete steps that will be taken by you to put your design that you 
thought of in previous phase into transformation.
Create a document around it and share the same for assessment.
Consider experimenting with ensemble methods or hyperparameter tuning to optimize the model's performance.